copy validation stuff to system.text
