+++ Valigator.AspNetCore.Text.Json/ValigatorJsonExceptionFilter.cs
@@ -1,12 +1,10 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
 using System;
 using System.Threading.Tasks;
-using Valigator.Newtonsoft.Json;
+using Valigator.Text.Json;
 
-namespace Valigator.AspNetCore.Newtonsoft.Json
+namespace Valigator.AspNetCore.Text.Json
 {
 	public class ValigatorJsonExceptionFilter : ExceptionFilterAttribute
 	{
