+++ Valigator.Text.Json/ValigatorJsonException.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Text.Json;
+
+namespace Valigator.Text.Json
+{
+	public class ValigatorJsonException : Exception
+	{
+		public ValigatorJsonException(JsonException jsonException) : base(jsonException.Message, jsonException) { }
+	}
+}
